MoveToLocation has been moved to Navigation class.
